+++ source/components/Autocomplete.js
@@ -12,7 +12,7 @@ import { GlobalListeners } from './HOC/GlobalListeners';
 import { withTheme } from './HOC/withTheme';
 
 // internal utility functions
-import { composeTheme, addThemeId } from '../utils/themes';
+import { composeTheme, addThemeId, didThemePropsChange } from '../utils/themes';
 import { composeFunctions } from '../utils/props';
 
 import { IDENTIFIERS } from '../themes/API';
@@ -106,6 +106,10 @@ class AutocompleteBase extends Component<Props, State> {
         context.ROOT_THEME_API
       )
     };
+  }
+
+  componentWillReceiveProps(nextProps: Props) {
+    didThemePropsChange(this.props, nextProps, this.setState.bind(this));
   }
 
   clear = () => this._removeOptions();