+++ source/components/Input.js
@@ -11,7 +11,7 @@ import { isString, flow } from 'lodash';
 import { withTheme } from './HOC/withTheme';
 
 // internal utility functions
-import { composeTheme, addThemeId } from '../utils/themes';
+import { composeTheme, addThemeId, didThemePropsChange } from '../utils/themes';
 
 // import constants
 import { IDENTIFIERS } from '../themes/API';
@@ -83,6 +83,10 @@ class InputBase extends Component<Props, State> {
 
   componentDidMount() {
     if (this.props.autoFocus) this.focus();
+  }
+
+  componentWillReceiveProps(nextProps: Props) {
+    didThemePropsChange(this.props, nextProps, this.setState.bind(this));
   }
 
   onChange = (event: SyntheticInputEvent<Element<'input'>>) => {