+++ source/themes/API/index.js
@@ -5,7 +5,10 @@ import { BUTTON_THEME_API } from './button';
 import { CHECKBOX_THEME_API } from './checkbox';
 import { FLEX_THEME_API } from './flex';
 import { FORM_FIELD_THEME_API } from './formfield';
+import { GRID_THEME_API } from './grid';
+import { GUTTER_THEME_API } from './gutter';
 import { HEADER_THEME_API } from './header';
+import { INFINITE_SCROLL_THEME_API } from './infinitescroll';
 import { INPUT_THEME_API } from './input';
 import { LOADING_SPINNER_API } from './loadingspinner';
 import { MODAL_THEME_API } from './modal';
@@ -25,7 +28,10 @@ export const IDENTIFIERS = {
   CHECKBOX: 'checkbox',
   FLEX: 'flex',
   FORM_FIELD: 'formfield',
+  GRID: 'grid',
+  GUTTER: 'gutter',
   HEADER: 'header',
+  INFINITE_SCROLL: 'infinitescroll',
   INPUT: 'input',
   LOADING_SPINNER: 'loadingspinner',
   MODAL: 'modal',
@@ -46,7 +52,10 @@ export const ROOT_THEME_API = {
   [IDENTIFIERS.CHECKBOX]: CHECKBOX_THEME_API,
   [IDENTIFIERS.FLEX]: FLEX_THEME_API,
   [IDENTIFIERS.FORM_FIELD]: FORM_FIELD_THEME_API,
+  [IDENTIFIERS.GRID]: GRID_THEME_API,
+  [IDENTIFIERS.GUTTER]: GUTTER_THEME_API,
   [IDENTIFIERS.HEADER]: HEADER_THEME_API,
+  [IDENTIFIERS.INFINITE_SCROLL]: INFINITE_SCROLL_THEME_API,
   [IDENTIFIERS.INPUT]: INPUT_THEME_API,
   [IDENTIFIERS.LOADING_SPINNER]: LOADING_SPINNER_API,
   [IDENTIFIERS.MODAL]: MODAL_THEME_API,