+++ source/themes/simple/index.js
@@ -9,7 +9,10 @@ import SimpleButton from './SimpleButton.scss';
 import SimpleCheckbox from './SimpleCheckbox.scss';
 import SimpleFlex from './SimpleFlex.scss';
 import SimpleFormField from './SimpleFormField.scss';
+import SimpleGrid from './SimpleGrid.scss';
+import SimpleGutter from './SimpleGutter.scss';
 import SimpleHeader from './SimpleHeader.scss';
+import SimpleInfiniteScroll from './SimpleInfiniteScroll.scss';
 import SimpleInput from './SimpleInput.scss';
 import SimpleLoadingSpinner from './SimpleLoadingSpinner.scss';
 import SimpleModal from './SimpleModal.scss';
@@ -33,7 +36,10 @@ export const SimpleTheme = {
   [IDENTIFIERS.CHECKBOX]: SimpleCheckbox,
   [IDENTIFIERS.FLEX]: SimpleFlex,
   [IDENTIFIERS.FORM_FIELD]: SimpleFormField,
+  [IDENTIFIERS.GRID]: SimpleGrid,
+  [IDENTIFIERS.GUTTER]: SimpleGutter,
   [IDENTIFIERS.HEADER]: SimpleHeader,
+  [IDENTIFIERS.INFINITE_SCROLL]: SimpleInfiniteScroll,
   [IDENTIFIERS.INPUT]: SimpleInput,
   [IDENTIFIERS.LOADING_SPINNER]: SimpleLoadingSpinner,
   [IDENTIFIERS.MODAL]: SimpleModal,