+++ stories/Layout.stories.js
@@ -7,15 +7,23 @@ import { storiesOf } from '@storybook/react';
 // components
 import { Flex } from '../source/components/layout/Flex';
 import { FlexItem } from '../source/components/layout/FlexItem';
+import { Grid } from '../source/components/layout/Grid';
+import { GridItem } from '../source/components/layout/GridItem';
+import { Gutter } from '../source/components/layout/Gutter';
+import { Header } from '../source/components/Header';
+
+// skins
+import { HeaderSkin } from '../source/skins/simple/HeaderSkin';
 
 // styles && themeOverrides
 import styles from './Layout.stories.scss';
 import customFlex from './theme-overrides/customFlex.scss';
+import customGrid from './theme-overrides/customGrid.scss';
 
 // decorator
 import { decorateWithSimpleTheme } from './helpers/theming';
 
-const { wrapper } = styles;
+const { wrapper, header, main, aside, footer, tomato, boxDark, boxLight, boxPastel } = styles;
 
 storiesOf('Layout', module)
 
@@ -127,4 +135,149 @@ storiesOf('Layout', module)
       <FlexItem flex={4}>flex = 4</FlexItem>
       <FlexItem flex={5}>flex = 5</FlexItem>
     </Flex>
-  ));
+  ))
+
+  .add('Grid - simple', () => (
+    <Gutter padding={60}>
+      <Grid columns="75% 25%" rows="10vw 30vw 10vw" gap={30}>
+        <div className={header}>header</div>
+        <div className={main}>main</div>
+        <div className={aside}>aside</div>
+        <div className={footer}>footer</div>
+      </Grid>
+    </Gutter>
+  ))
+
+  .add('Grid - even', () => (
+    <Gutter padding={30}>
+      <Grid
+        center
+        className={tomato}
+        columns="repeat(auto-fill, minmax(200px, 1fr))"
+        autoRows="minmax(150px, auto)"
+        gap="1em"
+      >
+        <div>1</div>
+        <div>2</div>
+        <div>3</div>
+        <div>4</div>
+        <div>5</div>
+        <div>6</div>
+        <div>7</div>
+        <div>8</div>
+        <div>9</div>
+        <div>10</div>
+        <div>11</div>
+        <div>12</div>
+        <div>13</div>
+        <div>14</div>
+        <div>15</div>
+      </Grid>
+    </Gutter>
+  ))
+
+  .add('Grid.Item - column & row', () => (
+    <Gutter padding="25vh 25vw">
+      <Grid
+        className={boxPastel}
+        columns="200px 200px 200px"
+        rows="125px 125px 125px"
+        gap={40}
+      >
+        <GridItem column="1 / 3" row="1">
+          1
+        </GridItem>
+        <GridItem column="3" row="1 / 3">
+          2
+        </GridItem>
+        <GridItem column="1" row="2">
+          3
+        </GridItem>
+        <GridItem column="2" row="2">
+          4
+        </GridItem>
+      </Grid>
+    </Gutter>
+  ))
+
+  .add('Grid.Item - columnStart/End rowStart/End', () => (
+    <Gutter padding="25vh 25vw">
+      <Grid className={boxLight} columns="repeat(3, 200px)" rows="repeat(3, 125px)" gap={10}>
+        <GridItem columnStart={2} columnEnd={3} rowStart={1}>
+          1
+        </GridItem>
+        <GridItem columnStart={3} rowStart={1} rowEnd={2}>
+          2
+        </GridItem>
+        <GridItem columnStart={1} rowStart={1} rowEnd={3}>
+          3
+        </GridItem>
+        <GridItem columnStart={2} columnEnd={4} rowStart={2}>
+          4
+        </GridItem>
+      </Grid>
+    </Gutter>
+  ))
+
+  .add('Grid - template & templateAreas', () => {
+    const templateAreas = [
+      'header header header',
+      'sidebar content content',
+      'sidebar content content',
+      'footer footer footer'
+    ];
+    return (
+      <Gutter padding="25vh 20vw">
+        <Grid
+          className={boxDark}
+          template="repeat(4, 1fr) / repeat(3, 1fr)"
+          templateAreas={templateAreas}
+          gap={10}
+        >
+          <GridItem gridArea="content">content</GridItem>
+          <GridItem gridArea="sidebar">sidebar</GridItem>
+          <GridItem gridArea="header">header</GridItem>
+          <GridItem gridArea="footer">footer</GridItem>
+        </Grid>
+      </Gutter>
+    );
+  })
+
+  .add('Grid - themeOverrides', () => {
+    const templateAreas = [
+      'sidebar header header header',
+      'sidebar content content aside',
+      'sidebar content content aside',
+      'sidebar footer footer footer'
+    ];
+    return (
+      <Gutter padding="25vh 20vw">
+        <Grid
+          themeOverrides={customGrid}
+          template="repeat(4, 1fr) / repeat(4, 1fr)"
+          templateAreas={templateAreas}
+          gap={10}
+        >
+          <GridItem gridArea="content">
+            <Header h2 left skin={HeaderSkin}>content</Header>
+          </GridItem>
+
+          <GridItem gridArea="sidebar">
+            <Header h2 left skin={HeaderSkin}>sidebar</Header>
+          </GridItem>
+
+          <GridItem gridArea="header">
+            <Header h2 left skin={HeaderSkin}>header</Header>
+          </GridItem>
+
+          <GridItem gridArea="footer">
+            <Header h2 left skin={HeaderSkin}>footer</Header>
+          </GridItem>
+
+          <GridItem gridArea="aside">
+            <Header h2 left skin={HeaderSkin}>aside</Header>
+          </GridItem>
+        </Grid>
+      </Gutter>
+    );
+  });